Separate systems into payloads and satellites
