+++ payload_designer/systems/__init__.py
@@ -1,0 +1,1 @@
+from . import payloads, satellites
